Migrate plugins on startup of redmine
Redmine has a rake task to iterate all plugins and perform migration steps. The reference can be found [here](https://www.redmine.org/projects/redmine/wiki/Plugins).

This will make it easier to integrate Plugins in the docker container. The user only has to mount a host folder with all the required plugins to /usr/src/redmine/plugins. With this rake task in place, all steps for making the plugins available are taken, though there may still be dependencies for the plugin that have to be installed first